- Replace `colog` with `env_logger` for logging. - Change `tick_count` type in `Config` from `u64` to `u8`. - Refactor `PacketSender` trait to use `enqueue` instead of `send`. - Modify `Server` struct to use `Arc<Mutex<ServerPacketSender>>`

+++ src/common/config.rs
@@ -6,7 +6,7 @@ use std::{env, fs};
 pub struct Config {
     pub db_uri: String,
     pub db_name: String,
-    pub tick_count: u64,
+    pub tick_count: u8,
 }
 
 impl Config {
